# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Micro-donation charity platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type CaseStatus = (typeof CaseStatus)[keyof typeof CaseStatus];

export const CaseStatus = {
  active: "active",
  funded: "funded",
  closed: "closed",
} as const;

export type CaseUrgencyLevel =
  (typeof CaseUrgencyLevel)[keyof typeof CaseUrgencyLevel];

export const CaseUrgencyLevel = {
  critical: "critical",
  high: "high",
  medium: "medium",
} as const;

export interface Case {
  id: number;
  title: string;
  description: string;
  patientName: string;
  patientAge: number;
  hospital: string;
  targetAmount: number;
  collectedAmount: number;
  sharePrice: number;
  totalShares: number;
  soldShares: number;
  status: CaseStatus;
  urgencyLevel: CaseUrgencyLevel;
  /** @nullable */
  imageUrl: string | null;
  /** @nullable */
  medicalReportUrl: string | null;
  createdAt: string;
  updatedAt: string;
}

export type CreateCaseBodyUrgencyLevel =
  (typeof CreateCaseBodyUrgencyLevel)[keyof typeof CreateCaseBodyUrgencyLevel];

export const CreateCaseBodyUrgencyLevel = {
  critical: "critical",
  high: "high",
  medium: "medium",
} as const;

export interface CreateCaseBody {
  title: string;
  description: string;
  patientName: string;
  patientAge: number;
  hospital: string;
  targetAmount: number;
  sharePrice: number;
  urgencyLevel: CreateCaseBodyUrgencyLevel;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  medicalReportUrl?: string | null;
}

export type UpdateCaseBodyUrgencyLevel =
  (typeof UpdateCaseBodyUrgencyLevel)[keyof typeof UpdateCaseBodyUrgencyLevel];

export const UpdateCaseBodyUrgencyLevel = {
  critical: "critical",
  high: "high",
  medium: "medium",
} as const;

export type UpdateCaseBodyStatus =
  (typeof UpdateCaseBodyStatus)[keyof typeof UpdateCaseBodyStatus];

export const UpdateCaseBodyStatus = {
  active: "active",
  funded: "funded",
  closed: "closed",
} as const;

export interface UpdateCaseBody {
  title?: string;
  description?: string;
  patientName?: string;
  patientAge?: number;
  hospital?: string;
  targetAmount?: number;
  sharePrice?: number;
  urgencyLevel?: UpdateCaseBodyUrgencyLevel;
  status?: UpdateCaseBodyStatus;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  medicalReportUrl?: string | null;
}

export interface DonateBody {
  donorName: string;
  shares: number;
  coverFees?: boolean;
  anonymous?: boolean;
}

export interface Donation {
  id: number;
  caseId: number;
  donorName: string;
  shares: number;
  amount: number;
  coverFees: boolean;
  anonymous: boolean;
  createdAt: string;
}

export type VoteStatus = (typeof VoteStatus)[keyof typeof VoteStatus];

export const VoteStatus = {
  open: "open",
  closed: "closed",
} as const;

/**
 * @nullable
 */
export type VoteResult = (typeof VoteResult)[keyof typeof VoteResult] | null;

export const VoteResult = {
  approved: "approved",
  rejected: "rejected",
} as const;

export interface Vote {
  id: number;
  caseId: number;
  title: string;
  description: string;
  expense: number;
  yesCount: number;
  noCount: number;
  status: VoteStatus;
  /** @nullable */
  result: VoteResult;
  createdAt: string;
  /** @nullable */
  closedAt: string | null;
}

export interface CreateVoteBody {
  caseId: number;
  title: string;
  description: string;
  expense: number;
}

export type CastVoteBodyChoice =
  (typeof CastVoteBodyChoice)[keyof typeof CastVoteBodyChoice];

export const CastVoteBodyChoice = {
  yes: "yes",
  no: "no",
} as const;

export interface CastVoteBody {
  choice: CastVoteBodyChoice;
  voterName: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  super_admin: "super_admin",
  sub_admin: "sub_admin",
  moderator: "moderator",
  donor: "donor",
} as const;

export interface User {
  id: number;
  name: string;
  email: string;
  role: UserRole;
  createdAt: string;
}

export type CreateUserBodyRole =
  (typeof CreateUserBodyRole)[keyof typeof CreateUserBodyRole];

export const CreateUserBodyRole = {
  super_admin: "super_admin",
  sub_admin: "sub_admin",
  moderator: "moderator",
  donor: "donor",
} as const;

export interface CreateUserBody {
  name: string;
  email: string;
  role: CreateUserBodyRole;
}

export type UpdateUserBodyRole =
  (typeof UpdateUserBodyRole)[keyof typeof UpdateUserBodyRole];

export const UpdateUserBodyRole = {
  super_admin: "super_admin",
  sub_admin: "sub_admin",
  moderator: "moderator",
  donor: "donor",
} as const;

export interface UpdateUserBody {
  name?: string;
  email?: string;
  role?: UpdateUserBodyRole;
}

export interface StatsOverview {
  totalDonations: number;
  totalCases: number;
  activeCases: number;
  fundedCases: number;
  totalDonors: number;
  totalSharesSold: number;
  averageFundingPercent: number;
}

export type ActivityItemType =
  (typeof ActivityItemType)[keyof typeof ActivityItemType];

export const ActivityItemType = {
  donation: "donation",
  case_funded: "case_funded",
  case_created: "case_created",
  vote_result: "vote_result",
} as const;

export interface ActivityItem {
  id: number;
  type: ActivityItemType;
  /** @nullable */
  donorName: string | null;
  caseTitle: string;
  caseId: number;
  /** @nullable */
  amount: number | null;
  /** @nullable */
  shares: number | null;
  createdAt: string;
}

export type ListCasesParams = {
  status?: ListCasesStatus;
  limit?: number;
  offset?: number;
};

export type ListCasesStatus =
  (typeof ListCasesStatus)[keyof typeof ListCasesStatus];

export const ListCasesStatus = {
  active: "active",
  funded: "funded",
  closed: "closed",
} as const;

export type ListDonationsParams = {
  caseId?: number;
  limit?: number;
  offset?: number;
};

export type ListVotesParams = {
  caseId?: number;
  status?: ListVotesStatus;
};

export type ListVotesStatus =
  (typeof ListVotesStatus)[keyof typeof ListVotesStatus];

export const ListVotesStatus = {
  open: "open",
  closed: "closed",
} as const;

export type GetRecentActivityParams = {
  limit?: number;
};