# lib/api-zod/src/generated/types/activityItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Micro-donation charity platform API
 * OpenAPI spec version: 0.1.0
 */
import type { ActivityItemType } from "./activityItemType";

export interface ActivityItem {
  id: number;
  type: ActivityItemType;
  /** @nullable */
  donorName: string | null;
  caseTitle: string;
  caseId: number;
  /** @nullable */
  amount: number | null;
  /** @nullable */
  shares: number | null;
  createdAt: string;
}