# lib/api-zod/src/generated/types/case.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Micro-donation charity platform API
 * OpenAPI spec version: 0.1.0
 */
import type { CaseStatus } from "./caseStatus";
import type { CaseUrgencyLevel } from "./caseUrgencyLevel";

export interface Case {
  id: number;
  title: string;
  description: string;
  patientName: string;
  patientAge: number;
  hospital: string;
  targetAmount: number;
  collectedAmount: number;
  sharePrice: number;
  totalShares: number;
  soldShares: number;
  status: CaseStatus;
  urgencyLevel: CaseUrgencyLevel;
  /** @nullable */
  imageUrl: string | null;
  /** @nullable */
  medicalReportUrl: string | null;
  createdAt: string;
  updatedAt: string;
}